requirements.txt
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
matplotlib==3.10.9
seaborn==0.13.2
tqdm==4.67.3
PyYAML==6.0.3